load real users in network page
